+++ zoom/_assets/standard_apps/home/app.py
@@ -49,6 +49,10 @@ def app(request):
         .zoom-app-icon .fa {
             font-size: 2em;
         }
+        .zoom-app-icon .bi {
+            font-size: 2rem;
+            line-height: 2rem;
+        }
     """
 
     if len(request.route) > 1 or request.data: