+++ zoom/packages.py
@@ -30,6 +30,11 @@ default_packages = {
             'jquery',
         ]
     },
+    "bootstrap-icons": {
+        "styles": [
+            "//cdn.jsdelivr.net/npm/bootstrap-icons@1.7.2/font/bootstrap-icons.css"
+        ]
+    },
     'd3': {
         'libs': [
             'https://cdnjs.cloudflare.com/ajax/libs/d3/3.5.17/d3.min.js',